requirements.txt
pygame==2.5.2
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
mapbox-vector-tile==2.2.0
pyserial==3.5